Made Child tracking indirect.
This way, no extra notifications are required when chunks load in/out.  It is up to the Node itself on how it wants to treat loaded/unloaded chunks.  However, on a BlockBreak, it will force load the chunks its parent and children are in, for the purposes of forcing them to load their data and remove the existing node.  However, as these nodes use indirection, it won't result in a chain of chunks loading.