Switch to Task/Worker Interfaces in distributed package.
Most of the heavy lifting will be done in concrete ITask implementations.  ITasks are also the only thing expected to have to be saved/loaded.  Worker -> Task assignments do not persist through save/load, as workers will actively seek new places to work as chunks unload.  Only thing to worry about might be mass chunk load/unload spam as workers get remigrated to workplaces that have not yet unloaded, but are about to be.

Might need to rearchitect the responsibility chain to accomodate, as well as incorporate indirect-ness, despite it all being in-memory anyway.  Event-driven may not work as due to above, and an active poll through pointers may be more accomodating of mass dropout.