TileFrame building progress is now server-side, and saves.
TileFrame rendering box needs to increase on the controller, based on size of the multiblock otherwise the inside stops rendering.
There appears to be a bug when the multiblock completes, where not always the frames won't be replaced, instead, only the controller is, and it's set to air, not an actual block.
Also NEED to refactor in-progress rendering to the associated IMultiblockRenderer.