Rendering fixups on Grasping Vines.
Fixed weird entity offsets on rendering.  Special-cased the player himself so vine wasn't floating.
Made grabbedSet actually look for and remove the entity.  Not sure why it wasn't throwing errors as that was a blatant violation of the typing system.
Entities that were grabbed that move away actually get moved away.  Moved this above creative player check because otherwise a creative player would be 'grabbed' and would render as 'grabbed' regardless of distance.
Reduced grab range.