Added some color to vines.
BasePos doesn't seem to persist through save/load.  This causes nullptrs on block break after a save/load.  Additionally, breaking a machine after placing it doesn't reactivate the 4-block machine in the build menu.  This is probably due to the 2nd block being 'leaked', and so the machine only sees 3 empty spaces. (1, 3, 4)