Continued Refactor of TileCyberBase.  [Still Broken]
Removed all rendering requirements from TileCyberBase.  Machine tracking should be handled much nicer, now.

Issues:
Tracking Machines-In-Progress.  Need to think of some clever way of tracking both machines-in-progress, but also the amount of space they're going to take up.  Probably would be best to update CyberMachines to allow for dynamic sizing based on location, but the issue with that is rendering in menus and such where there is no 'location' to poll information from.
Otherwise, need special MachineMapping cases to handle it.