Refactor out Graph behaviors into interfaces.  Soon to follow with default tracking implementations.
Started work on IPowerZiggurat infrastructure, which lead to the refactor.  Need to refactor the PowerUtils to work on said refactored-out architecture as well, as there is no reason it should work ONLY for IPowerZiggurats as well (except for how it determines end-goals and other things, which could be done with lambda functions.)

The idea being, I've done this structure so much already, might as well pull it out to a clear interface.  It just makes things clearer without having to constantly redo behavior.