Started work on ItemRepository, an IndexedInventory chest of 9x6 slots. Implementation of it would hopefully have searchbar integration for ore ids and item names (potentially itemIDs) as well, player-side.
This also implements the as-of-yet unused NaniteManager registration (potentially to be renamed) which means this hooks into and remembers a parent Nanite Hive, and is also remembered by said Hive.  This is for future inventory searches.

Also fixed up some behavior in IndexedInventory, such as removing unnecessary outside reference to the itemstack in the slot to be removed in RemoveItemStack call.  This just introduced the possibility for an outside call to give the wrong itemstack.  Instead, just check it ourselves and remove it from teh cache as necessary.

Refactor nanite/Trait to nanite/trait.